:memo: :: rename module and pakage
