+++ feature/root/src/main/java/khs/onmi/root/MainActivity.kt
@@ -12,7 +12,7 @@ import khs.onmi.core.designsystem.theme.ONMITheme
 import khs.onmi.enterinformation.navigation.enterInformationNavGraph
 import khs.onmi.main.navigation.mainNavGraph
 import khs.onmi.navigation.ONMINavRoutes
-import khs.onmi.setup.navigation.setupNavGraph
+import khs.onmi.setting.navigation.settingNavGraph
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
@@ -38,7 +38,7 @@ class MainActivity : ComponentActivity() {
                             navController = navController
                         )
 
-                        setupNavGraph(navController = navController)
+                        settingNavGraph(navController = navController)
                     }
                 }
             }
