+++ feature/setting/src/main/AndroidManifest.xml
