+++ feature/setting/src/main/java/khs/onmi/setting/screen/MainRoute.kt
@@ -1,4 +1,4 @@
-package khs.onmi.setup.screen
+package khs.onmi.setting.screen
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
