+++ feature/setting/src/test/java/khs/onmi/setting/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package khs.onmi.setup
+package khs.onmi.setting
 
 import org.junit.Test
