+++ settings.gradle.kts
@@ -20,7 +20,7 @@ include(":navigation")
 include(":feature:enterinformation")
 include(":feature:main")
 include(":feature:notice")
-include(":feature:setup")
+include(":feature:setting")
 include(":feature:allergies")
 include(":feature:timetable")
 include(":feature:tutorials")
